+++ example/App.js
@@ -6,109 +6,95 @@
  * @flow
  */
 
-import React from 'react';
-import {
-  SafeAreaView,
-  StyleSheet,
-  ScrollView,
-  View,
-  Text,
-  StatusBar,
-} from 'react-native';
+import React, {Component} from 'react';
+import {Platform, StyleSheet, Text, View, Switch, YellowBox} from 'react-native';
+import {ChannelList, BroadcastModalView, BroadcastVideo} from 'boxcast-sdk-react-native';
 
-import {
-  Header,
-  LearnMoreLinks,
-  Colors,
-  DebugInstructions,
-  ReloadInstructions,
-} from 'react-native/Libraries/NewAppScreen';
+// TODO: look up your own channel IDs
+const MY_BOXCAST_CHANNEL_ID = 'lbkvcqkzmxyhzwzsbj6w';
 
-const App: () => React$Node = () => {
-  return (
-    <>
-      <StatusBar barStyle="dark-content" />
-      <SafeAreaView>
-        <ScrollView
-          contentInsetAdjustmentBehavior="automatic"
-          style={styles.scrollView}>
-          <Header />
-          {global.HermesInternal == null ? null : (
-            <View style={styles.engine}>
-              <Text style={styles.footer}>Engine: Hermes</Text>
-            </View>
-          )}
-          <View style={styles.body}>
-            <View style={styles.sectionContainer}>
-              <Text style={styles.sectionTitle}>Step One</Text>
-              <Text style={styles.sectionDescription}>
-                Edit <Text style={styles.highlight}>App.js</Text> to change this
-                screen and then come back to see your edits.
-              </Text>
-            </View>
-            <View style={styles.sectionContainer}>
-              <Text style={styles.sectionTitle}>See Your Changes</Text>
-              <Text style={styles.sectionDescription}>
-                <ReloadInstructions />
-              </Text>
-            </View>
-            <View style={styles.sectionContainer}>
-              <Text style={styles.sectionTitle}>Debug</Text>
-              <Text style={styles.sectionDescription}>
-                <DebugInstructions />
-              </Text>
-            </View>
-            <View style={styles.sectionContainer}>
-              <Text style={styles.sectionTitle}>Learn More</Text>
-              <Text style={styles.sectionDescription}>
-                Read the docs to discover what to do next:
-              </Text>
-            </View>
-            <LearnMoreLinks />
-          </View>
-        </ScrollView>
-      </SafeAreaView>
-    </>
-  );
-};
-
-const styles = StyleSheet.create({
-  scrollView: {
-    backgroundColor: Colors.lighter,
-  },
-  engine: {
-    position: 'absolute',
-    right: 0,
-  },
-  body: {
-    backgroundColor: Colors.white,
-  },
-  sectionContainer: {
-    marginTop: 32,
-    paddingHorizontal: 24,
-  },
-  sectionTitle: {
-    fontSize: 24,
-    fontWeight: '600',
-    color: Colors.black,
-  },
-  sectionDescription: {
-    marginTop: 8,
-    fontSize: 18,
-    fontWeight: '400',
-    color: Colors.dark,
-  },
-  highlight: {
-    fontWeight: '700',
-  },
-  footer: {
-    color: Colors.dark,
-    fontSize: 12,
-    fontWeight: '600',
-    padding: 4,
-    paddingRight: 12,
-    textAlign: 'right',
-  },
+BroadcastVideo.configureAnalytics({
+  host: 'My BoxCast RN App',
+  os: Platform.OS,
+  browser_name: 'React Native',
+  browser_version: '1.0',
+  player_version: 'boxcast-sdk-react-native v1.0'
 });
 
-export default App;
+YellowBox.ignoreWarnings([
+  'Accessing view manager configs',
+  'Invalid image url',
+  'source.uri should not be',
+  'Task orphaned',
+  'Remote debugger is in a background tab',
+]);
+
+type Props = {};
+export default class App extends Component<Props> {
+  state = {
+    broadcast: null,
+    dockable: true,
+  };
+
+  render() {
+    return (
+      <View style={styles.container}>
+
+        <View style={{marginBottom: 15, width: '100%', alignItems: 'center'}}>
+          <Text>ChannelList Horizontal Mode</Text>
+          <View style={{height: 150, width: '100%', borderWidth: 1, borderColor: '#ff0000'}}>
+            <ChannelList channelId={MY_BOXCAST_CHANNEL_ID}
+                        query={'timeframe:relevant timeframe:next'}
+                        sort={'-starts_at'}
+                        pageSize={10}
+                        onSelectBroadcast={(broadcast) => this.showBroadcast(broadcast)}
+                        horizontal={true} />
+          </View>
+        </View>
+
+        <View style={{marginBottom: 15, width: '100%', alignItems: 'center'}}>
+          <Text>ChannelList Vertical Mode</Text>
+          <View style={{height: 300, width: '75%', borderWidth: 1, borderColor: '#ff0000'}}>
+            <ChannelList channelId={MY_BOXCAST_CHANNEL_ID}
+                        query={'timeframe:relevant timeframe:next'}
+                        sort={'-starts_at'}
+                        pageSize={10}
+                        onSelectBroadcast={(broadcast) => this.showBroadcast(broadcast)}
+                        horizontal={false} />
+          </View>
+        </View>
+
+        <View style={{flexDirection: 'row', alignItems: 'center'}}>
+          <Switch onValueChange={(dockable) => this.setState({dockable})} value={this.state.dockable} />
+          <Text style={{marginLeft: 5}}>Allow video docking?</Text>
+        </View>
+
+        {this.state.broadcast &&
+          <BroadcastModalView
+              broadcast={this.state.broadcast}
+              dockable={this.state.dockable}
+              onDismiss={() => this.closeBroadcast()}
+              debug={true}
+          />
+        }
+      </View>
+    );
+  }
+
+  showBroadcast(broadcast) {
+    this.setState({broadcast});
+  }
+
+  closeBroadcast() {
+    this.setState({broadcast: null});
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#F5FCFF',
+  },
+});