+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGELOG
 
+### v0.2.0 on (unreleased)
+
+* Upgrade to support React Native v0.61
+* Upgrade to react-native-video v5.0
+
 ### v0.1.0 on 2019-06-26
 
-* Initial release
+* Initial release with support for React Native v0.59